+++ src/m365/spo/commands/group/group-member-add.ts
@@ -24,13 +24,13 @@ interface Options extends GlobalOptions {
   email?: string;
 }
 
-class SpoGroupUserAddCommand extends SpoCommand {
+class SpoGroupMemberAddCommand extends SpoCommand {
   public get name(): string {
-    return commands.GROUP_USER_ADD;
+    return commands.GROUP_MEMBER_ADD;
   }
 
   public get description(): string {
-    return 'Add a user or multiple users to SharePoint Group';
+    return 'Add members to a SharePoint Group';
   }
 
   public defaultProperties(): string[] | undefined {
@@ -221,4 +221,4 @@ class SpoGroupUserAddCommand extends SpoCommand {
   }
 }
 
-module.exports = new SpoGroupUserAddCommand();
+module.exports = new SpoGroupMemberAddCommand();