+++ docs/docs/about/release-notes.md
@@ -96,6 +96,7 @@
 - Included response output for yammer commands [#3928](https://github.com/pnp/cli-microsoft365/issues/3928)
 - Updates GitHub actions to Node@16 [#3870](https://github.com/pnp/cli-microsoft365/issues/3870)
 - Updates docs for 'spo eventreceiver remove' command [#4050](https://github.com/pnp/cli-microsoft365/issues/4050)
+- Updates 'spo group user <verb>' to 'spo group member <verb>' [#4087](https://github.com/pnp/cli-microsoft365/issues/4087)
 
 ## [v5.9.0](https://github.com/pnp/cli-microsoft365/releases/tag/v5.9.0)
 
@@ -579,7 +580,7 @@
 
 **SharePoint:**
 
-- [spo group user remove](../cmd/spo/group/group-user-remove.md) - removes the specified user from a SharePoint group [#2642](https://github.com/pnp/cli-microsoft365/issues/2642)
+- [spo group user remove](../cmd/spo/group/group-member-remove.md) - removes the specified user from a SharePoint group [#2642](https://github.com/pnp/cli-microsoft365/issues/2642)
 
 **Teams:**
 
@@ -772,7 +773,7 @@
 
 **SharePoint:**
 
-- [spo group user list](../cmd/spo/group/group-user-list.md) - list members of a SharePoint Group [#2640](https://github.com/pnp/cli-microsoft365/issues/2640)
+- [spo group user list](../cmd/spo/group/group-member-list.md) - list members of a SharePoint Group [#2640](https://github.com/pnp/cli-microsoft365/issues/2640)
 
 **Teams:**
 
@@ -813,7 +814,7 @@
 
 **SharePoint:**
 
-- [spo group user add](../cmd/spo/group/group-user-add.md) - add a user or multiple users to SharePoint Group [#2483](https://github.com/pnp/cli-microsoft365/issues/2483)
+- [spo group user add](../cmd/spo/group/group-member-add.md) - add a user or multiple users to SharePoint Group [#2483](https://github.com/pnp/cli-microsoft365/issues/2483)
 - [spo site apppermission remove](../cmd/spo/site/site-apppermission-remove.md) - removes a specific application permission from a site [#2209](https://github.com/pnp/cli-microsoft365/issues/2209)
 - [spo app instance list](../cmd/spo/app/app-instance-list.md) - retrieves apps installed in a site [#2025](https://github.com/pnp/cli-microsoft365/issues/2209)